Move stats delete helper out of route
