+++ src/app/api/user/[userId]/stats/route.ts
@@ -1,4 +1,3 @@
-import { TZDate } from "@date-fns/tz";
 import { NextRequest, NextResponse } from "next/server";
 import { Prisma } from "@prisma/client";
 import { auth } from "@/auth";
@@ -23,61 +22,12 @@ import {
   type TotalsAccumulator,
 } from "@/app/api/user/[userId]/stats/types";
 import { type StatsCollection } from "@/app/_stats/types";
+import { getDeleteDateBounds } from "./delete-date-bounds";
 
 function parseDailyStatsDay(day: DailyStatsRow["day"]): Date {
   return day instanceof Date ? day : new Date(day);
 }
 
-function parseLocalDateParts(date: string): {
-  year: number;
-  monthIndex: number;
-  day: number;
-} {
-  const match = /^(\d{4})-(\d{2})-(\d{2})$/.exec(date);
-  if (!match) {
-    return { year: NaN, monthIndex: NaN, day: NaN };
-  }
-
-  return {
-    year: Number(match[1]),
-    monthIndex: Number(match[2]) - 1,
-    day: Number(match[3]),
-  };
-}
-
-function localDateParamToUtc(date: string, timezone: string): Date {
-  const { year, monthIndex, day } = parseLocalDateParts(date);
-  return new Date(new TZDate(year, monthIndex, day, timezone).getTime());
-}
-
-export function getDeleteDateBounds(
-  startDate: string,
-  endDate: string,
-  timezone: string | null
-): { startDateTime: Date; endDateTime: Date } {
-  if (timezone) {
-    const startDateTime = localDateParamToUtc(startDate, timezone);
-    const endParts = parseLocalDateParts(endDate);
-    const endExclusiveDateTime = new Date(
-      new TZDate(
-        endParts.year,
-        endParts.monthIndex,
-        endParts.day + 1,
-        timezone
-      ).getTime()
-    );
-
-    return {
-      startDateTime,
-      endDateTime: new Date(endExclusiveDateTime.getTime() - 1),
-    };
-  }
-
-  return {
-    startDateTime: new Date(`${startDate}T00:00:00.000Z`),
-    endDateTime: new Date(`${endDate}T23:59:59.999Z`),
-  };
-}
 
 function getDaysInUtcMonth(date: Date): number {
   return new Date(
